Signal1DTools into Signal1D
